refactor: UserPosts 상태 page로 이동
