+++ src/components/pages/UserPage/index.js
@@ -1,23 +1,50 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { useLocation } from "react-router";
+import { endLoading, startLoading } from "../../../modules/loading";
+import { getPostsAPI } from "../../../utils/API";
 import Footer from "../../organisms/Footer";
 import NavBar from "../../organisms/NavBar";
 import UserPosts from "../../organisms/UserPosts";
 import NavAndFooter from "../../templates/NavAndFooter";
 
-const UserPage = (props) => {
-  // console.log("userpage");
+const UserPage = () => {
+  const { pathname } = useLocation();
+  const username = pathname.substr(1);
 
-  // 이것도 useEffect로?
-  // useEffect(() => {
-  //   //
-  // },[props.location.pathname]);
+  const dispatch = useDispatch();
+  const [error, setError] = useState("");
+  const [posts, setPosts] = useState([]);
 
-  const username = props.location.pathname.substr(1); // ❓ 아 모르겠다... 작동 원리를 모르겠다. route와 component의... 작동은 되는데 이게 좋은 방법일까?
+  useEffect(() => {
+    const doGetPostsAPI = async (username) => {
+      dispatch(startLoading());
+      setError("");
+      try {
+        const posts = await getPostsAPI(username);
+        dispatch(endLoading());
+        setError("");
+        setPosts([...posts]);
+      } catch (e) {
+        dispatch(endLoading());
+        setError(e.message);
+      }
+    };
+
+    doGetPostsAPI(username);
+  }, [username, dispatch]);
+
+  const UserPostsProps = {
+    error,
+    setError,
+    posts,
+    setPosts,
+  };
 
   return (
     <NavAndFooter
       nav={<NavBar />}
-      context={<UserPosts username={username} />}
+      context={<UserPosts {...UserPostsProps} />}
       footer={<Footer />}
     />
   );
